backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.0
drf-yasg==1.21.10
gunicorn==23.0.0
groq==0.18.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
whitenoise==6.9.0
